+++ install/dotfiles.sh
@@ -7,17 +7,28 @@ for entry in dotfiles/*; do
 	# Any existing files will be renamed .bak
 	if [ -f "$entry" ]; then
 		target=~/."$(basename "$entry")"
-		[ -e "$target" ] && mv "$target" "$target.bak"
-		ln -s "$(pwd)/$entry" "$target"
+
+		if [ -e "$target" ] && [ "$(readlink "$target")" != "$(pwd)/$entry" ]; then
+			mv "$target" "$target.bak"
+		fi
+
+		if [ ! -e "$target" ]; then
+			ln -s "$(pwd)/$entry" "$target"
+		fi
 	fi
 
 	# Link all directories in ~/.config/
 	# Any existing directories will be renamed .bak
 	if [ -d "$entry" ]; then
 		target=~/.config/"$(basename "$entry")"
-		[ -e "$target" ] && mv "$target" "$target.bak"
-		ln -s "$(pwd)/$entry" "$target"
+
+		if [ -e "$target" ] && [ "$(readlink "$target")" != "$(pwd)/$entry" ]; then
+			mv "$target" "$target.bak"
+		fi
+
+		if [ ! -e "$target" ]; then
+			ln -s "$(pwd)/$entry" "$target"
+		fi
 	fi
 done
-
 unset entry
